@@ -1,14 +1,6 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
-import "./css/style.css";
-import { BootstrapVue } from "bootstrap-vue";
-import "bootstrap/dist/css/bootstrap.css";
-import "bootstrap-vue/dist/bootstrap-vue.css";
+import router from "./router";
+import store from "./store";
 
-Vue.use(BootstrapVue);
-
-Vue.config.productionTip = false;
-
-new Vue({
-  render: (h) => h(App),
-}).$mount("#app");
+createApp(App).use(store).use(router).mount("#app");
